requirements.txt
opencv-python
numpy
torch
torchvision
facenet-pytorch
scikit-learn
qdrant-client
streamlit
Pillow
